ADD MerchantViewModel
