+++ WalletLess/Views/merchant/Merchants.swift
@@ -2,51 +2,10 @@
 
 import SwiftUI
 import Foundation
-import FirebaseFirestore
-import Firebase
 import RealmSwift
 import CodeScanner
 import MessageUI
-import FirebaseStorage
 import Combine
-
-class MerchantsViewModel: ObservableObject {
-    
-    @Published var merchants = [MerchantDto]()
-    let databaseName: String = "MKD"
-    
-    private let storage = Storage.storage().reference()
-    
-    func fetchDataIfNeeded() {
-        let database = Database.database().reference().child(databaseName)
-        
-        database.observe(.value) { snapshot in
-            
-            guard let dict = snapshot.value as? [String:Any] else { return }
-            
-            dict.forEach { card in
-                let card = card.value as? [String: Any]
-                
-                let name = card?["cardName"] as? String ?? ""
-                let locations = card?["locations"] as? [String: [String: Any]]
-                
-                self.storage.child("MKD/\(name).png").downloadURL(completion: { url, error in
-                    guard let url = url, error == nil else { return }
-                    guard let imageURL = URL(string: url.absoluteString) else { return }
-                    
-                    URLSession.shared.dataTask(with: imageURL, completionHandler: { data, _, error in
-                        guard let data = data, error == nil else { return }
-                        
-                        DispatchQueue.main.async {
-                            self.merchants.append(MerchantDto(name: name, downloadedImage: data, locations: locations))
-                        }
-                        
-                    }).resume()
-                })
-            }
-        }
-    }
-}
 
 struct Merchants: View {
     
@@ -135,24 +94,6 @@ struct Merchants: View {
                                 .padding(.bottom, 10)
                                 
                             }
-                            // TODO: FKJ - Feature Show Gallery on camera view
-                            //                                HStack {
-                            //                                    Spacer()
-                            //                                    Button(action: {
-                            //                                        isGalleryPresented.toggle()
-                            //                                        requestGalleryPermission()
-                            //                                    }, label: {
-                            //                                        Image(systemName: "photo.on.rectangle")
-                            //                                            .font(.title2)
-                            //                                            .frame(width: 50, height: 50)
-                            //                                            .background(networkManger.isConnected ? Color.blue : Color.gray)
-                            //                                            .clipShape(Circle())
-                            //                                            .foregroundColor(.white)
-                            //                                    })
-                            //                                    .padding(.trailing)
-                            //                                    .padding(.bottom, 10)
-                            //                                    .shadow(radius: 2)
-                            //                                }
                         }
                     }
                 }
@@ -201,35 +142,6 @@ struct Merchants: View {
         }
     }
     
-    // MARK: -
-    
-    // TODO: FKJ - Feature Show Gallery on camera view
-    //    private func requestGalleryPermission() {
-    //        PHPhotoLibrary.requestAuthorization(for: .readWrite) { [self] (status) in
-    //            DispatchQueue.main.async { [self] in
-    //                showUI(for: status)
-    //            }
-    //        }
-    //    }
-    //
-    //    private func showUI(for status: PHAuthorizationStatus) {
-    //
-    //        switch status {
-    //        case .authorized:
-    //            print("Authorized")
-    //        case .limited:
-    //            print("Limited")
-    //        case .restricted:
-    //            print("Restricted")
-    //        case .denied:
-    //            print("Denied")
-    //        case .notDetermined:
-    //            break
-    //        @unknown default:
-    //            break
-    //        }
-    //    }
-    
     private func presentMailSheet() {
         if MFMailComposeViewController.canSendMail() {
             self.isPresentingMailView = true
@@ -269,34 +181,4 @@ struct Merchants_Previews: PreviewProvider {
     static var previews: some View {
         Merchants(tabSelection: Binding.constant(2))
     }
-}
-
-struct MerchantsViewModifier: ViewModifier {
-    
-    @State private var didLoad = false
-    
-    private let action: (() -> Void)?
-    
-    init(perform action: (() -> Void)?) {
-        self.action = action
-    }
-    
-    func body(content: Content) -> some View {
-        content.onAppear {
-            if !didLoad {
-                didLoad = true
-                action?()
-            }
-        }
-    }
-    
-}
-
-extension View {
-    
-    // Custom modifier for checking if the view did load.
-    func viewDidLoad(perform action: (() -> Void)?) -> some View {
-        modifier(MerchantsViewModifier(perform: action))
-    }
-    
 }
